complete HW 2

+++ hw2/hw2/PercolationStats.java
@@ -33,7 +33,7 @@ public class PercolationStats {
             col = StdRandom.uniform(gridSize);
             p.open(row, col);
         }
-        return p.numberOfOpenSites() / (gridSize * gridSize);
+        return (double)p.numberOfOpenSites() / (double)(gridSize * gridSize);
     }
 
     public double mean() {
